+++ internal/infrastructure/helpers/http_helper.go
@@ -1,0 +1,7 @@
+package helpers
+
+import "net/http"
+
+func IsResponseSuccessful(response *http.Response) bool {
+	return response.StatusCode >= http.StatusOK && response.StatusCode < http.StatusMultipleChoices
+}