+++ CHANGELOG.md
@@ -16,6 +16,17 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+### Added
+
+- added JSON handling to perform better operations with the response
+- added state importing feature with Base64 encoding
+
+### Changed
+
+- corrected basic testing with basic checking with 3 cases
+- corrected documentation to have examples in the official page
+- corrected the structure to be more readable using DDD
+
 ## [1.0.0] - 2024-08-24
 
 ### Changed